Converted copy board to push/pop move
