+++ agent.py
@@ -30,16 +30,16 @@ class Agent :
 		best_move = None
 
 		for move in moves :
-			next_state = board.copy(stack=False)
-			next_state.push(move)
+			board.push(move)
 			features = []
 			for f in extractors :
-				res = f.extract(next_state, color)
+				res = f.extract(board, color)
 				features += res
 			quality = self.net.activate(tuple(features))[0]
 			if quality > max_quality :
 				max_quality = quality
 				best_move = move
+			board.pop()
 		
 		assert best_move != None
 		return best_move
